Move RandomNumberGenerator objects into fillWithRandomIslands()

+++ shm/RandomNumberGenerator.hpp
@@ -6,11 +6,11 @@ class RandomNumberGenerator {
 public:
     RandomNumberGenerator()
         : generator_{std::random_device{}()}, distribution_(0, 10) {}
-    int nextRandomNumber() {
+    size_t nextRandomNumber() {
         return distribution_(generator_);
     }
 
 private:
     std::mt19937 generator_;
-    std::uniform_int_distribution<int> distribution_;
+    std::uniform_int_distribution<size_t> distribution_;
 };
